Bind custom model function to HuggingFaceHandlerService instance
This commit changes the way the custom model_fn is bound to the instance
of the HuggingFaceHandlerService so the custom model_fn has access to
the context available to the handler.

This will allow us to use that context, for example `self.device` to
ensure we properly set the device when creating a custom pipeline.